Update table body generation and move to ui crate.
